+++ lib/src/mt_instance.h
@@ -10,14 +10,14 @@
 int mt_instance_init(struct mtl_main_impl* impl, struct mtl_init_params* p);
 int mt_instance_uinit(struct mtl_main_impl* impl);
 
-int mt_instance_get_lcore(struct mtl_main_impl* impl, unsigned int lcore_id);
-int mt_instance_put_lcore(struct mtl_main_impl* impl, unsigned int lcore_id);
+int mt_instance_get_lcore(struct mtl_main_impl* impl, uint16_t lcore_id);
+int mt_instance_put_lcore(struct mtl_main_impl* impl, uint16_t lcore_id);
 int mt_instance_request_xsks_map_fd(struct mtl_main_impl* impl, unsigned int ifindex);
 int mt_instance_update_udp_dp_filter(struct mtl_main_impl* impl, unsigned int ifindex,
                                      uint16_t dst_port, bool add);
 int mt_instance_get_queue(struct mtl_main_impl* impl, unsigned int ifindex,
-                          unsigned int queue_id);
+                          uint16_t queue_id);
 int mt_instance_put_queue(struct mtl_main_impl* impl, unsigned int ifindex,
-                          unsigned int queue_id);
+                          uint16_t queue_id);
 
 #endif
